+++ src/main/java/open/commons/utils/CollectionUtils.java
@@ -31,7 +31,6 @@ import java.util.AbstractCollection;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
-import java.util.Collections;
 import java.util.Comparator;
 import java.util.Enumeration;
 import java.util.HashMap;
@@ -1743,7 +1742,7 @@ public class CollectionUtils {
      */
     public static <E> List<E> subCollection(List<E> list, int begin) {
         if (begin > list.size() - 1) {
-            return Collections.emptyList();
+            return new ArrayList<E>();
         }
 
         List<E> newList = new ArrayList<E>();
@@ -1768,7 +1767,7 @@ public class CollectionUtils {
      */
     public static <E> List<E> subCollection(List<E> list, int begin, int end) {
         if (begin > list.size() - 1 || begin >= end || end < 0) {
-            return Collections.emptyList();
+            return new ArrayList<E>();
         }
 
         List<E> newList = new ArrayList<E>();